Move prototype cross simd wrapper structs to CrossSIMD.h
